requirements.txt
asn1crypto==1.5.1
cffi==1.17.1
coincurve==19.0.1
colorama==0.4.6
coverage==7.6.8
green==3.3.0
lxml==5.3.0
protobuf==3.19.6
pycparser==2.22
pycryptodome==3.19.1
sjcl==0.2.1
Unidecode==1.3.8
